pyproject.toml
[project]
name = "de-server"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "playwright>=1.54.0",
    "pytest-playwright>=0.7.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "openai>=1.0.0",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.3",
    "pillow>=11.3.0",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pillow>=11.3.0",
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 160
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
